# packages/sdk/openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "openpatent",
    "description": "openpatent api",
    "version": "1.0.0"
  },
  "paths": {
...
    "/tui/control/response": {
      "post": {
        "operationId": "tui.control.response",
        "parameters": [
          {
            "in": "query",
            "name": "directory",
            "schema": {
              "type": "string"
            }
          }
        ],
        "summary": "Submit TUI response",
        "description": "Submit a response to the TUI request queue to complete a pending request.",
        "responses": {
          "200": {
            "description": "Response submitted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "boolean"
                }
              }
            }
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {}
            }
          }
        },
        "x-codeSamples": [
          {
            "lang": "js",
            "source": "import { createopenpatentClient } from \"@openpatent-ai/sdk\n\nconst client = createopenpatentClient()\nawait client.tui.control.response({\n  ...\n})"
          }
        ]
      }
    },
...
  },
...
}